requirements.txt
Flask==3.1.3
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
PyMySQL==1.1.2
python-dotenv==1.2.1
Werkzeug==3.1.6